Update spec to use loadTime
We now associate image requests to elements: <img>, SVG's <image>, and <video> can have a direct association, and Elements also have a list due to background images. With this association, we can compute the loadTime and no longer need responseEnd.

The "Get response end time algorithm" is kept since it is still used by LargestContentfulPaint, but should be removed soon.